# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Source API",
        "description": "API for Source - Figma Design Audit Platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/projects/{projectId}/screens/import": {
            "post": {
                "tags": [
                    "Screens"
                ],
                "summary": "Import screens from Figma",
                "description": "Import multiple Figma frames as screens for a project",
                "operationId": "3275d012ef6e21e1a87478a01cec9096",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "frame_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "1:23",
                                            "1:24"
                                        ]
                                    },
                                    "figma_access_token": {
                                        "type": "string",
                                        "example": "figd_abc123..."
                                    },
                                    "figma_file_key": {
                                        "type": "string",
                                        "example": "abc123def456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Screens imported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Screen"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/projects/{projectId}/screens": {
            "get": {
                "tags": [
                    "Screens"
                ],
                "summary": "Get project screens",
                "description": "Get all screens for a project with optional search",
                "operationId": "ca560b90875ce3ccfa1d2c9d1012852b",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search screens by name or description",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of project screens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Screen"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/projects/{projectId}/screens/{screenId}": {
            "get": {
                "tags": [
                    "Screens"
                ],
                "summary": "Get screen by ID",
                "description": "Get a specific screen by its ID",
                "operationId": "533fc8d111634a8b1cd641e336e82cca",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "screenId",
                        "in": "path",
                        "description": "Screen ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Screen data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Screen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Screen not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Screens"
                ],
                "summary": "Update screen",
                "description": "Update a screen by its ID",
                "operationId": "517a8d1e6943b5faf0ac2097963e441c",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "screenId",
                        "in": "path",
                        "description": "Screen ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "section_name": {
                                        "type": "string",
                                        "example": "Updated Screen Name",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Screen updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Screen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Screen not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Screens"
                ],
                "summary": "Delete screen",
                "description": "Delete a screen by its ID",
                "operationId": "92a44433e6d3764d2292fd502c05b261",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "screenId",
                        "in": "path",
                        "description": "Screen ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Screen deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Screen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Screen not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Screen": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "section_name": {
                        "type": "string",
                        "example": "Login Screen",
                        "nullable": true
                    },
                    "figma_node_id": {
                        "type": "string",
                        "example": "1:23",
                        "nullable": true
                    },
                    "figma_file_key": {
                        "type": "string",
                        "example": "abc123def456",
                        "nullable": true
                    },
                    "figma_svg_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://figma.com/svg/...",
                        "nullable": true
                    },
                    "figma_node_name": {
                        "type": "string",
                        "example": "Login",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "User login screen",
                        "nullable": true
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}